remove old periodic dependency
